zine: make test failures more debuggable

+++ internal/tool/zine/zine_test.go
@@ -14,7 +14,10 @@ func TestFullRender(t *testing.T) {
 	if err != nil {
 		t.Fatalf("couldn't create test: %s", err)
 	}
-	defer os.RemoveAll(d)
+	if os.Getenv("LM_KEEP") == "" {
+		t.Log("set LM_KEEP and the directory will not be cleaned up")
+		defer os.RemoveAll(d)
+	}
 
 	if err := render([]string{"render", "-static", "./testdata", "-root", "./testdata", "-out", d}); err != nil {
 		t.Errorf("Rendered produced unexpected error: %s", err)
@@ -23,6 +26,7 @@ func TestFullRender(t *testing.T) {
 	b, err := ioutil.ReadFile(filepath.Join(d, "cats", "index.html"))
 	if err != nil {
 		t.Errorf("Couldn't read output: %s", err)
+		return
 	}
 
 	testutil.Equal(t, `<p>This is the header!</p>
